update actions and reducers for meeting notifications
