+++ src/components/Header/Header.jsx
@@ -58,11 +58,8 @@ import PopUpBar from 'components/PopUpBar';
 import './Header.css';
 import hasPermission, { cantUpdateDevAdminDetails } from '../../utils/permissions';
 import { fetchTaskEditSuggestions } from 'components/TaskEditSuggestions/thunks';
-import {
-  getUnreadUserNotifications,
-  getUnreadMeetingNotification,
-  markMeetingNotificationAsRead,
-} from '../../actions/notificationAction';
+import { getUnreadUserNotifications } from '../../actions/notificationAction';
+import { getUnreadMeetingNotification, markMeetingNotificationAsRead, } from '../../actions/meetingNotificationAction';
 import { toast } from 'react-toastify';
 import NotificationCard from '../Notification/notificationCard';
 import DarkModeButton from './DarkModeButton';
@@ -159,11 +156,13 @@ export function Header(props) {
   const [meetingModalMessage, setMeetingModalMessage] = useState('');
 
   // const unreadNotifications = props.unreadNotifications; // List of unread notifications
-  const { allUserProfiles, unreadNotifications, unreadMeetingNotifications } = props;
+  const { allUserProfiles, notification, unreadNotifications, unreadMeetingNotifications } = props;
   // const userUnreadMeetings = useMemo(() => {
   //   if (!unreadMeetingNotifications || !userId) return [];
   //   return unreadMeetingNotifications.filter(meeting => meeting.recipient === userId);
   // }, [unreadMeetingNotifications, userId]);
+  console.log('test', notification);
+  console.log('unreadMeetingNotifications', unreadMeetingNotifications, typeof unreadMeetingNotifications);
   const userUnreadMeetings = unreadMeetingNotifications.filter(meeting => meeting.recipient === userId);
   const dispatch = useDispatch();
   const history = useHistory();
@@ -622,7 +621,8 @@ const mapStateToProps = state => ({
   role: state.role,
   notification: state.notification,
   unreadNotifications: state.notification.unreadNotifications,
-  unreadMeetingNotifications: state.notification.unreadMeetingNotifications,
+  // unreadMeetingNotifications: state.notification.unreadMeetingNotifications,
+  unreadMeetingNotifications: state.meetingNotification.unreadMeetingNotifications,
   allUserProfiles: state.allUserProfiles.userProfiles,
   darkMode: state.theme.darkMode,
 });
