+++ src/reducers/notificationReducer.js
@@ -1,10 +1,10 @@
 import * as actionTypes from '../constants/notification';
-import * as meetingActions from '../constants/meetings';
+// import * as meetingActions from '../constants/meetings';
 
 const initialState = {
   notifications: [], // all notifications. This is used in the notification history for admin/owner.
   unreadNotifications: [], // unread notifications. This is to store all unread notification for a user.
-  unreadMeetingNotifications: [],
+  // unreadMeetingNotifications: [],
   sentNotifications: [],
   loading: false,
   error: null,
@@ -19,8 +19,8 @@ const notificationReducer = (state = initialState, action) => {
     case actionTypes.CREATE_NOTIFICATION_REQUEST:
     case actionTypes.DELETE_NOTIFICATION_REQUEST:
     case actionTypes.MARK_NOTIFICATION_AS_READ_REQUEST:
-    case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_BEGIN:
-    case meetingActions.MARK_MEETING_AS_READ_REQUEST:
+    // case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_BEGIN:
+    // case meetingActions.MARK_MEETING_AS_READ_REQUEST:
       return {
         ...state,
         loading: true,
@@ -35,8 +35,8 @@ const notificationReducer = (state = initialState, action) => {
     case actionTypes.CREATE_NOTIFICATION_FAILURE:
     case actionTypes.DELETE_NOTIFICATION_FAILURE:
     case actionTypes.MARK_NOTIFICATION_AS_READ_FAILURE:
-    case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_FAILURE:
-    case meetingActions.MARK_MEETING_AS_READ_FAILURE:
+    // case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_FAILURE:
+    // case meetingActions.MARK_MEETING_AS_READ_FAILURE:
       return {
         ...state,
         loading: false,
@@ -65,14 +65,16 @@ const notificationReducer = (state = initialState, action) => {
         error: null,
       };
 
-    case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_SUCCESS:
-      return {
-        ...state,
-        // unreadNotifications: [...state.unreadNotifications, ...action.payload],
-        unreadMeetingNotifications: action.payload,
-        loading: false,
-        error: null,
-      }
+    // Huijie
+    // case meetingActions.FETCH_UNREAD_UPCOMING_MEETING_SUCCESS:
+    //   return {
+    //     ...state,
+    //     // unreadNotifications: [...state.unreadNotifications, ...action.payload],
+    //     unreadMeetingNotifications: action.payload,
+    //     loading: false,
+    //     error: null,
+    //   }
+    // Huijie
 
     // case actionTypes.FETCH_SENT_NOTIFICATIONS_SUCCESS:
     //   return {
@@ -109,17 +111,18 @@ const notificationReducer = (state = initialState, action) => {
         error: null,
       };
     
-    case meetingActions.MARK_MEETING_AS_READ_SUCCESS:
-      const { unreadMeetingNotifications } = state;
-      const newUnreadMeetingNotifications = unreadMeetingNotifications.filter((notification) => !(notification.meetingId === action.payload.meetingId && notification.recipient === action.payload.recipient));
-      return {
-        ...state,
-        // remove the meeting notification from unreadNotifications
-        unreadMeetingNotifications: newUnreadMeetingNotifications,
-        loading: false,
-        error: null,
-      };
-
+    // Huijie
+    // case meetingActions.MARK_MEETING_AS_READ_SUCCESS:
+    //   const { unreadMeetingNotifications } = state;
+    //   const newUnreadMeetingNotifications = unreadMeetingNotifications.filter((notification) => !(notification.meetingId === action.payload.meetingId && notification.recipient === action.payload.recipient));
+    //   return {
+    //     ...state,
+    //     // remove the meeting notification from unreadNotifications
+    //     unreadMeetingNotifications: newUnreadMeetingNotifications,
+    //     loading: false,
+    //     error: null,
+    //   };
+    // Huije
     
     default:
       return state;
